resumise/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
google-generativeai>=0.5.2
PyPDF2==3.0.1
beautifulsoup4==4.12.3
sentence-transformers==2.3.1
torch>=2.0.0
python-multipart==0.0.9
